chore: index.spec.ts has been split into their own files
